+++ Specs/Scene/ModelSpec.js
@@ -4,7 +4,9 @@ defineSuite([
         'Core/Cartesian2',
         'Core/Cartesian3',
         'Core/Cartesian4',
+        'Core/combine',
         'Core/defaultValue',
+        'Core/defined',
         'Core/FeatureDetection',
         'Core/HeadingPitchRange',
         'Core/JulianDate',
@@ -15,6 +17,7 @@ defineSuite([
         'Core/PrimitiveType',
         'Core/Transforms',
         'Renderer/RenderState',
+        'Renderer/ShaderSource',
         'Renderer/WebGLConstants',
         'Scene/ModelAnimationLoop',
         'Specs/createScene',
@@ -25,7 +28,9 @@ defineSuite([
         Cartesian2,
         Cartesian3,
         Cartesian4,
+        combine,
         defaultValue,
+        defined,
         FeatureDetection,
         HeadingPitchRange,
         JulianDate,
@@ -36,6 +41,7 @@ defineSuite([
         PrimitiveType,
         Transforms,
         RenderState,
+        ShaderSource,
         WebGLConstants,
         ModelAnimationLoop,
         createScene,
@@ -119,20 +125,33 @@ defineSuite([
     }
 
     function loadModel(url, options) {
-        options = defaultValue(options, {});
-
-        var model = primitives.add(Model.fromGltf({
+        options = combine(options, {
+            modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
             url : url,
-            modelMatrix : defaultValue(options.modelMatrix, Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0))),
-            show : false,
-            scale : options.scale,
-            minimumPixelSize : options.minimumPixelSize,
-            maximumScale : options.maximumScale,
-            id : url,        // for picking tests
-            asynchronous : options.asynchronous,
-            releaseGltfJson : options.releaseGltfJson,
-            cacheKey : options.cacheKey
-        }));
+            id : url,
+            show : false
+        });
+
+        var model = primitives.add(Model.fromGltf(options));
+        addZoomTo(model);
+
+        return pollToPromise(function() {
+            // Render scene to progressively load the model
+            scene.renderForSpecs();
+            return model.ready;
+        }, { timeout: 10000 }).then(function() {
+            return model;
+        });
+    }
+
+    function loadModelJson(gltf, options) {
+        options = combine(options, {
+            modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
+            gltf : gltf,
+            show : false
+        });
+
+        var model = primitives.add(new Model(options));
         addZoomTo(model);
 
         return pollToPromise(function() {
@@ -149,8 +168,10 @@ defineSuite([
         expect(scene.renderForSpecs()).toEqual([0, 0, 0, 255]);
         model.show = true;
         model.zoomTo();
-        expect(scene.renderForSpecs()).not.toEqual([0, 0, 0, 255]);
+        var pixelColor = scene.renderForSpecs();
+        expect(pixelColor).not.toEqual([0, 0, 0, 255]);
         model.show = false;
+        return pixelColor;
     }
 
     it('fromGltf throws without options', function() {
@@ -198,36 +219,17 @@ defineSuite([
 
     it('renders from glTF', function() {
         // Simulate using procedural glTF as opposed to loading it from a file
-        var model = primitives.add(new Model({
-            gltf : texturedBoxModel.gltf,
-            modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-            show : false
-        }));
-        addZoomTo(model);
-
-        return pollToPromise(function() {
-            // Render scene to progressively load the model
-            scene.renderForSpecs();
-            return model.ready;
-        }, { timeout: 10000 }).then(function() {
+        return loadModelJson(texturedBoxModel.gltf).then(function(model) {
             verifyRender(model);
             primitives.remove(model);
         });
     });
 
     it('Applies the right render state', function() {
-        // Simulate using procedural glTF as opposed to loading it from a file
-        var model = primitives.add(new Model({
-            gltf : texturedBoxModel.gltf
-        }));
-
         spyOn(RenderState, 'fromCache').and.callThrough();
 
-        return pollToPromise(function() {
-            // Render scene to progressively load the model
-            scene.renderForSpecs();
-            return model.ready;
-        }, { timeout : 10000 }).then(function() {
+        // Simulate using procedural glTF as opposed to loading it from a file
+        return loadModelJson(texturedBoxModel.gltf).then(function(model) {
             var rs = {
                 frontFace : WebGLConstants.CCW,
                 cull : {
@@ -363,7 +365,7 @@ defineSuite([
         expect(texturedBoxModel.getMesh('name-of-mesh-that-does-not-exist')).not.toBeDefined();
     });
 
-    it('getMesh returns returns a mesh', function() {
+    it('getMesh returns a mesh', function() {
         var mesh = texturedBoxModel.getMesh('Mesh');
         expect(mesh).toBeDefined();
         expect(mesh.name).toEqual('Mesh');
@@ -388,7 +390,7 @@ defineSuite([
         expect(texturedBoxModel.getMaterial('name-of-material-that-does-not-exist')).not.toBeDefined();
     });
 
-    it('getMaterial returns returns a material', function() {
+    it('getMaterial returns a material', function() {
         var material = texturedBoxModel.getMaterial('Texture');
         expect(material).toBeDefined();
         expect(material.name).toEqual('Texture');
@@ -547,20 +549,7 @@ defineSuite([
 
     it('loads a model with the CESIUM_binary_glTF extension as an ArrayBuffer using new Model', function() {
         return loadArrayBuffer(texturedBoxBinaryUrl).then(function(arrayBuffer) {
-            var model = primitives.add(new Model({
-                gltf : arrayBuffer,
-                modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-                show : false
-            }));
-            addZoomTo(model);
-
-            return pollToPromise(function() {
-                // Render scene to progressively load the model
-                scene.renderForSpecs();
-                return model.ready;
-            }, {
-                timeout : 10000
-            }).then(function() {
+            return loadModelJson(arrayBuffer).then(function(model) {
                 verifyRender(model);
                 primitives.remove(model);
             });
@@ -569,20 +558,7 @@ defineSuite([
 
     it('loads a model with the CESIUM_binary_glTF extension as an Uint8Array using new Model', function() {
         return loadArrayBuffer(texturedBoxBinaryUrl).then(function(arrayBuffer) {
-            var model = primitives.add(new Model({
-                gltf : new Uint8Array(arrayBuffer),
-                modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-                show : false
-            }));
-            addZoomTo(model);
-
-            return pollToPromise(function() {
-                // Render scene to progressively load the model
-                scene.renderForSpecs();
-                return model.ready;
-            }, {
-                timeout : 10000
-            }).then(function() {
+            return loadModelJson(new Uint8Array(arrayBuffer)).then(function(model) {
                 verifyRender(model);
                 primitives.remove(model);
             });
@@ -598,20 +574,7 @@ defineSuite([
 
     it('loads a model with the KHR_binary_glTF extension as an ArrayBuffer using new Model', function() {
         return loadArrayBuffer(texturedBoxKhrBinaryUrl).then(function(arrayBuffer) {
-            var model = primitives.add(new Model({
-                gltf : arrayBuffer,
-                modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-                show : false
-            }));
-            addZoomTo(model);
-
-            return pollToPromise(function() {
-                // Render scene to progressively load the model
-                scene.renderForSpecs();
-                return model.ready;
-            }, {
-                timeout : 10000
-            }).then(function() {
+            return loadModelJson(arrayBuffer).then(function(model) {
                 verifyRender(model);
                 primitives.remove(model);
             });
@@ -620,20 +583,7 @@ defineSuite([
 
     it('loads a model with the KHR_binary_glTF extension as an Uint8Array using new Model', function() {
         return loadArrayBuffer(texturedBoxKhrBinaryUrl).then(function(arrayBuffer) {
-            var model = primitives.add(new Model({
-                gltf : new Uint8Array(arrayBuffer),
-                modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-                show : false
-            }));
-            addZoomTo(model);
-
-            return pollToPromise(function() {
-                // Render scene to progressively load the model
-                scene.renderForSpecs();
-                return model.ready;
-            }, {
-                timeout : 10000
-            }).then(function() {
+            return loadModelJson(new Uint8Array(arrayBuffer)).then(function(model) {
                 verifyRender(model);
                 primitives.remove(model);
             });
@@ -1213,7 +1163,7 @@ defineSuite([
         });
     });
 
-    it('releaseGltfJson releases glTFJSON when constructed with fromGltf', function() {
+    it('releaseGltfJson releases glTF JSON when constructed with fromGltf', function() {
         return loadModel(boxUrl, {
             releaseGltfJson : true
         }).then(function(m) {
@@ -1226,21 +1176,11 @@ defineSuite([
     });
 
     it('releaseGltfJson releases glTF JSON when constructed with Model constructor function', function() {
-        var m = primitives.add(new Model({
-            gltf : texturedBoxModel.gltf,
-            modelMatrix : Transforms.eastNorthUpToFixedFrame(Cartesian3.fromDegrees(0.0, 0.0, 100.0)),
-            show : false,
+        return loadModelJson(texturedBoxModel.gltf, {
             releaseGltfJson : true,
             incrementallyLoadTextures : false,
             asynchronous : true
-        }));
-        addZoomTo(m);
-
-        return pollToPromise(function() {
-            // Render scene to progressively load the model
-            scene.renderForSpecs();
-            return m.ready;
-        }, { timeout : 10000 }).then(function() {
+        }).then(function(m) {
             expect(m.releaseGltfJson).toEqual(true);
             expect(m.gltf).not.toBeDefined();
 
@@ -1470,6 +1410,46 @@ defineSuite([
         });
     });
 
+    it('Loads with incrementallyLoadTextures set to true', function() {
+        return loadModelJson(texturedBoxModel.gltf, {
+            incrementallyLoadTextures : true
+        }).then(function(m) {
+            // Get the rendered color of the model before textures are loaded
+            var loadedColor = verifyRender(m);
+
+            pollToPromise(function() {
+                // Render scene to progressively load textures
+                scene.renderForSpecs();
+                // Textures have finished loading
+                return (m.pendingTextureLoads === 0);
+            }, { timeout : 10000 }).then(function() {
+                var finishedColor = verifyRender(m);
+                expect(finishedColor).not.toEqual(loadedColor);
+                primitives.remove(m);
+            });
+        });
+    });
+
+    it('Loads with incrementallyLoadTextures set to false', function() {
+        return loadModelJson(texturedBoxModel.gltf, {
+            incrementallyLoadTextures : false
+        }).then(function(m) {
+            // Get the rendered color of the model before textures are loaded
+            var loadedColor = verifyRender(m);
+
+            pollToPromise(function() {
+                // Render scene to progressively load textures (they should already be loaded)
+                scene.renderForSpecs();
+                // Textures have finished loading
+                return !defined(m._loadResources);
+            }, { timeout : 10000 }).then(function() {
+                var finishedColor = verifyRender(m);
+                expect(finishedColor).toEqual(loadedColor);
+                primitives.remove(m);
+            });
+        });
+    });
+
     it('loads a glTF with KHR_materials_common using the constant lighting model', function() {
         return loadModel(boxConstantUrl).then(function(m) {
             verifyRender(m);
@@ -1550,6 +1530,150 @@ defineSuite([
     it('loads a glTF with KHR_materials_common that has transparency', function() {
         return loadModel(boxTransparentUrl).then(function(m) {
             verifyRender(m);
+            primitives.remove(m);
+        });
+    });
+
+    it('loads with custom vertex attributes, vertexShader, fragmentShader, and uniform map', function() {
+        function vertexShaderLoaded(vs) {
+            var renamedSource = ShaderSource.replaceMain(vs, 'czm_old_main');
+            var newMain =
+                'attribute vec4 a_color;\n' +
+                'varying vec4 v_color;\n' +
+                'void main()\n' +
+                '{\n' +
+                '    czm_old_main();\n' +
+                '    v_color = a_color;\n' +
+                '}';
+            return renamedSource + '\n' + newMain;
+        }
+
+        function fragmentShaderLoaded(fs) {
+            fs = 'uniform float u_value;\n' +
+                 'varying vec4 v_color;\n' +
+                 'void main()\n' +
+                 '{\n' +
+                 '    gl_FragColor = u_value * v_color;\n' +
+                 '}';
+            return fs;
+        }
+
+        function uniformMapLoaded(uniformMap) {
+            return combine(uniformMap, {
+                u_value : function() {
+                    return 1.0;
+                }
+            });
+        }
+
+        var precreatedAttributes = {
+            a_color : {
+                index                  : 0, // updated in Model
+                componentsPerAttribute : 4,
+                value                  : [1.0, 1.0, 1.0, 1.0]
+            }
+        };
+
+        var options = {
+            precreatedAttributes : precreatedAttributes,
+            vertexShaderLoaded : vertexShaderLoaded,
+            fragmentShaderLoaded : fragmentShaderLoaded,
+            uniformMapLoaded : uniformMapLoaded
+        };
+
+        return loadModelJson(texturedBoxModel.gltf, options).then(function(model) {
+            var pixelColor = verifyRender(model);
+            expect(pixelColor).toEqual([255, 255, 255, 255]);
+            primitives.remove(model);
+        });
+    });
+
+    it('loads with custom pickFragmentShader and pickUniformMap', function() {
+        function pickFragmentShaderLoaded(fs) {
+            return ShaderSource.createPickFragmentShaderSource(fs, 'uniform');
+        }
+
+        var pickId = scene.context.createPickId({
+            custom : 'custom'
+        });
+
+        function pickUniformMapLoaded(uniformMap) {
+            return combine(uniformMap, {
+                czm_pickColor : function() {
+                    return pickId.color;
+                }
+            });
+        }
+
+        var options = {
+            pickFragmentShaderLoaded : pickFragmentShaderLoaded,
+            pickUniformMapLoaded : pickUniformMapLoaded
+        };
+
+        return loadModelJson(texturedBoxModel.gltf, options).then(function(model) {
+            model.show = true;
+            var pick = scene.pick(new Cartesian2(0, 0));
+            expect(pick.custom).toEqual('custom');
+            primitives.remove(model);
+        });
+    });
+
+    it('does not issue draw commands when ignoreCommands is true', function() {
+        return loadModel(texturedBoxUrl, {
+            ignoreCommands : true
+        }).then(function(m) {
+            expect(m.ready).toBe(true);
+            m.show = true;
+
+            m.zoomTo();
+            m.update(scene.frameState);
+            expect(scene.frameState.commandList.length).toEqual(0);
+
+            m.show = false;
+            primitives.remove(m);
+        });
+    });
+
+    it('does not issue draw commands when the model is out of view and cull is true', function() {
+        return loadModel(texturedBoxUrl, {
+            cull : true
+        }).then(function(m) {
+            expect(m.ready).toBe(true);
+            m.show = true;
+
+            // Look at the model
+            m.zoomTo();
+            scene.renderForSpecs();
+            expect(scene._frustumCommandsList.length).not.toEqual(0);
+
+            // Move the model out of view
+            m.modelMatrix = Matrix4.fromTranslation(new Cartesian3(10000000000.0, 0.0, 0.0));
+            scene.renderForSpecs();
+            expect(scene._frustumCommandsList.length).toEqual(0);
+
+            m.show = false;
+            primitives.remove(m);
+        });
+    });
+
+    it('issues draw commands when the model is out of view and cull is false', function() {
+        return loadModel(texturedBoxUrl, {
+            cull : false
+        }).then(function(m) {
+            expect(m.ready).toBe(true);
+            m.show = true;
+
+            // Look at the model
+            m.zoomTo();
+            scene.renderForSpecs();
+            expect(scene._frustumCommandsList.length).not.toEqual(0);
+
+            // Move the model out of view
+            m.modelMatrix = Matrix4.fromTranslation(new Cartesian3(10000000000.0, 0.0, 0.0));
+            scene.renderForSpecs();
+            expect(scene._frustumCommandsList.length).not.toEqual(0);
+
+            m.show = false;
             primitives.remove(m);
         });
     });